+++ src/Providers/WidgetServiceProvider.php
@@ -44,6 +44,7 @@ class WidgetServiceProvider extends ServiceProvider
     private array $builtInModifiers = [
         AppShellWidgetModifiers::UPPERCASE => Widgets\Modifiers\Uppercase::class,
         AppShellWidgetModifiers::LOWERCASE => Widgets\Modifiers\Lowercase::class,
+        AppShellWidgetModifiers::INTVAL => Widgets\Modifiers\Intval::class,
         AppShellWidgetModifiers::BOOL2TEXT => Widgets\Modifiers\Bool2Text::class,
         AppShellWidgetModifiers::TEXT_IF_EMPTY => Widgets\Modifiers\TextIfEmpty::class,
         AppShellWidgetModifiers::SHOW_DATETIME => Widgets\Modifiers\ShowDateTime::class,